Base Excel Sheet
Enter a login for lync into cell A1 and hit the button. The cell should
change color depending on the status. Right now the sheet checks for
away, offline, or online. THIS HAS NOT BEEN TESTED IN AN ACTUAL
ENVIRONMENT. I coded this from my memory of doing it where I could test
it, so it might not work.